Allowing "Relations Only" save format for TypeaheadList pickers
The "Relations Only" save format option was not available for
TypeaheadList pickers. The reason was because the selected options might
not be present in the selectable collection. To ensure that the selected
options are available, a new method called
getEditorDataItemsFilteredByIds() has been added to all datasources and
is called to populate the selected data items (only for TypeaheadList).