Added delay for the TypeaheadListPicker
With the existing controller, a new AJAX request will be made for each character entered in the search field. With this commit, I've added a delay, so the AJAX request is first fired 250 ms after the last entered character. This will minimize calls made to the server, and also give editors a better experience in the UI.