Fixed issue for nuPickers added to Membertypes
When a nuPicker is added to a custum MemberType and an instcance of that membertype is loaded through use of the memberservice i.e. Umbraco.Web.Composing.Current.UmbracoHelper.Member(id), the PickerPropertyValueConverter throws an error on line: 117.

The check assignedContentItem.Parent != null appears not to be a valid check for IMember eventhough it's successfully cast to IPublishedContent, the fix works around the check by getting the parent id from the Path property.